# Conflicts:
#	app/controllers/Application.scala
#	app/controllers/Datasets.scala
#	app/controllers/Spaces.scala
#	app/services/DI.scala
#	app/views/datasetList.scala.html
#	app/views/datasets/tile.scala.html
#	app/views/home.scala.html
#	app/views/main.scala.html
#	app/views/spaces/datasetsBySpace.scala.html
#	app/views/spaces/datasetsGrid.scala.html
#	app/views/spaces/space.scala.html
#	public/stylesheets/main.css

+++ app/services/mongodb/MongoDBFileService.scala
@@ -5,32 +5,39 @@ import services._
 import models._
 import com.mongodb.casbah.commons.MongoDBObject
 import java.text.SimpleDateFormat
-import _root_.util.{Parsers, License}
+
+import _root_.util.{License, Parsers, SearchUtils}
 
 import scala.collection.mutable.ListBuffer
 import Transformation.LidoToCidocConvertion
-import java.util.{Calendar, ArrayList}
+import java.util.{ArrayList, Calendar}
 import java.io._
+
 import org.apache.commons.io.FileUtils
 import org.json.JSONObject
-import play.api.libs.json.{Json, JsValue}
+import play.api.libs.json.{JsValue, Json}
 import com.mongodb.util.JSON
 import java.nio.file.{FileSystems, Files}
 import java.nio.file.attribute.BasicFileAttributes
+
 import collection.JavaConverters._
 import scala.collection.JavaConversions._
 import javax.inject.{Inject, Singleton}
+
 import com.mongodb.casbah.WriteConcern
 import play.api.Logger
+
 import scala.util.parsing.json.JSONArray
 import play.api.libs.json.JsArray
 import models.File
 import play.api.libs.json.JsObject
 import java.util.Date
+
 import com.novus.salat.dao.{ModelCompanion, SalatDAO}
 import MongoContext.context
 import play.api.Play._
 import com.mongodb.casbah.Imports._
+import models.FileStatus.FileStatus
 
 
 /**
@@ -63,6 +70,12 @@ class MongoDBFileService @Inject() (
     FileDAO.count(MongoDBObject())
   }
 
+  def statusCount(): Map[FileStatus, Long] = {
+    val results = FileDAO.dao.collection.aggregate(MongoDBObject("$group" ->
+      MongoDBObject("_id" -> "$status", "count" -> MongoDBObject("$sum" -> 1L))))
+    results.results.map(x => FileStatus.withName(x.getAsOrElse[String]("_id", FileStatus.UNKNOWN.toString)) -> x.getAsOrElse[Long]("count", 0L)).toMap
+  }
+
   def bytes(): Long = {
     val results = FileDAO.dao.collection.aggregate(MongoDBObject("$group" ->
       MongoDBObject("_id" -> "size", "total" -> MongoDBObject("$sum" -> "$length"))))
@@ -73,7 +86,7 @@ class MongoDBFileService @Inject() (
   }
 
   def save(file: File): Unit = {
-    FileDAO.save(file)
+    FileDAO.save(file, WriteConcern.Safe)
   }
 
   /**
@@ -99,7 +112,7 @@ class MongoDBFileService @Inject() (
       FileDAO.find("isIntermediate" $ne true).sort(order).limit(limit).toList
     } else {
       val sinceDate = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS").parse(date)
-      Logger.info("After " + sinceDate)
+      Logger.debug("After " + sinceDate)
       FileDAO.find($and("isIntermediate" $ne true, "uploadDate" $lt sinceDate)).sort(order).limit(limit).toList
     }
   }
@@ -114,7 +127,7 @@ class MongoDBFileService @Inject() (
     } else {
       order = MongoDBObject("uploadDate" -> 1)
       val sinceDate = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS").parse(date)
-      Logger.info("Before " + sinceDate)
+      Logger.debug("Before " + sinceDate)
       FileDAO.find($and("isIntermediate" $ne true, "uploadDate" $gt sinceDate)).sort(order).limit(limit).toList.reverse
     }
   }
@@ -174,8 +187,7 @@ class MongoDBFileService @Inject() (
   def save(inputStream: InputStream, filename: String, contentType: Option[String], author: MiniUser, showPreviews: String = "DatasetLevel"): Option[File] = {
     ByteStorageService.save(inputStream, FileDAO.COLLECTION) match {
       case Some(x) => {
-
-        val file = File(UUID.generate(), x._1, filename, author, new Date(), util.FileUtils.getContentType(filename, contentType), x._4, x._3, x._2, showPreviews = showPreviews, licenseData = License.fromAppConfig())
+        val file = File(UUID.generate(), x._1, filename, author, new Date(), util.FileUtils.getContentType(filename, contentType), x._3, x._2, showPreviews = showPreviews, licenseData = License.fromAppConfig())
         FileDAO.save(file)
         Some(file)
       }
@@ -202,47 +214,9 @@ class MongoDBFileService @Inject() (
   def index(id: UUID) {
     get(id) match {
       case Some(file) => {
-        var tagListBuffer = new ListBuffer[String]()
-
-        for (tag <- file.tags) {
-          tagListBuffer += tag.name
-        }
-
-        val tagsJson = new JSONArray(tagListBuffer.toList)
-
-        Logger.debug("tagStr=" + tagsJson);
-
-        val commentsByFile = for (comment <- comments.findCommentsByFileId(id)) yield {
-          comment.text
-        }
-        val commentJson = new JSONArray(commentsByFile)
-
-        Logger.debug("commentStr=" + commentJson.toString())
-
-        val usrMd = getUserMetadataJSON(id)
-        Logger.debug("usrmd=" + usrMd)
-
-        val techMd = getTechnicalMetadataJSON(id)
-        Logger.debug("techmd=" + techMd)
-
-        val xmlMd = getXMLMetadataJSON(id)
-        Logger.debug("xmlmd=" + xmlMd)
-
-        var fileDsId = ""
-        var fileDsName = ""
-
-        for (dataset <- datasets.findByFileId(file.id)) {
-          fileDsId = fileDsId + dataset.id.stringify + " %%% "
-          fileDsName = fileDsName + dataset.name + " %%% "
-        }
-        
-        val formatter = new SimpleDateFormat("dd/MM/yyyy")
-
         current.plugin[ElasticsearchPlugin].foreach {
-          _.index("data", "file", id,
-            List(("filename", file.filename), ("contentType", file.contentType),("author",file.author.fullName),("uploadDate",formatter.format(file.uploadDate)),("datasetId",fileDsId),("datasetName",fileDsName), ("tag", tagsJson.toString), ("comments", commentJson.toString), ("usermetadata", usrMd), ("technicalmetadata", techMd), ("xmlmetadata", xmlMd)))
+          _.index(file)
         }
-        
       }
       case None => Logger.error("File not found: " + id)
     }
@@ -304,7 +278,7 @@ class MongoDBFileService @Inject() (
         val theJSON = getUserMetadataJSON(id)
         val fileSep = System.getProperty("file.separator")
         val tmpDir = System.getProperty("java.io.tmpdir")
-        var resultDir = tmpDir + fileSep + "medici__rdfuploadtemporaryfiles" + fileSep + UUID.generate.stringify
+        var resultDir = tmpDir + fileSep + "clowder__rdfuploadtemporaryfiles" + fileSep + UUID.generate.stringify
         val resultDirFile = new java.io.File(resultDir)
         resultDirFile.mkdirs()
 
@@ -475,10 +449,15 @@ class MongoDBFileService @Inject() (
     }
   }
 
+  override def setStatus(id: UUID, status: FileStatus): Unit = {
+    FileDAO.dao.update(MongoDBObject("_id" -> new ObjectId(id.toString())), $set("status" -> status.toString))
+  }
+
+
   def listOutsideDataset(dataset_id: UUID): List[File] = {
     datasets.get(dataset_id) match{
       case Some(dataset) => {
-        val list = for (file <- FileDAO.findAll(); if(!isInDataset(file,dataset) && !file.isIntermediate.getOrElse(false))) yield file
+        val list = for (file <- FileDAO.findAll(); if(!isInDataset(file,dataset) && !file.isIntermediate)) yield file
         return list.toList
       }
       case None =>{
@@ -551,7 +530,10 @@ class MongoDBFileService @Inject() (
     }
   }
 
-  
+  /** Change the metadataCount field for a file */
+  def incrementMetadataCount(id: UUID, count: Long) = {
+    FileDAO.update(MongoDBObject("_id" -> new ObjectId(id.stringify)), $inc("metadataCount" -> count), false, false, WriteConcern.Safe)
+  }
   
   /**
    *  Add versus descriptors to the Versus.descriptors collection associated to a file
@@ -561,7 +543,7 @@ class MongoDBFileService @Inject() (
     val doc = JSON.parse(Json.stringify(json)).asInstanceOf[DBObject].toMap
               .asScala.asInstanceOf[scala.collection.mutable.Map[String,Any]].toMap
        VersusDAO.insert(new Versus(id,doc),WriteConcern.Safe)
-       Logger.info("--Added versus descriptors in json format received from versus to the metadata field --")
+       Logger.debug("--Added versus descriptors in json format received from versus to the metadata field --")
   }
  
 /**
@@ -694,7 +676,7 @@ class MongoDBFileService @Inject() (
   def removeFile(id: UUID){
     get(id) match{
       case Some(file) => {
-        if(file.isIntermediate.isEmpty){
+        if(!file.isIntermediate){
           val fileDatasets = datasets.findByFileId(file.id)
           for(fileDataset <- fileDatasets){
             datasets.removeFile(fileDataset.id, id)
@@ -705,20 +687,19 @@ class MongoDBFileService @Inject() (
             if(!file.thumbnail_id.isEmpty && !fileDataset.thumbnail_id.isEmpty){            
               if(file.thumbnail_id.get.equals(fileDataset.thumbnail_id.get)){ 
                 datasets.newThumbnail(fileDataset.id)
-                
                 	for(collectionId <- fileDataset.collections){
-		                          collections.get(collectionId) match{
-		                            case Some(collection) =>{		                              
-		                            	if(!collection.thumbnail_id.isEmpty){	                            	  
-		                            		if(collection.thumbnail_id.get.equals(fileDataset.thumbnail_id.get)){
-		                            			collections.createThumbnail(collection.id)
-		                            		}		                        
-		                            	}
-		                            }
-		                            case None=>Logger.debug(s"Could not find collection $collectionId") 
-		                          }
-		                        }		        	  
-		        }
+                    collections.get(collectionId) match{
+                      case Some(collection) =>{
+                        if(!collection.thumbnail_id.isEmpty){
+                          if(collection.thumbnail_id.get.equals(fileDataset.thumbnail_id.get)){
+                            collections.createThumbnail(collection.id)
+                          }
+                        }
+                      }
+                      case None=>Logger.debug(s"Could not find collection $collectionId")
+                    }
+                  }
+		          }
             }
                      
           }
@@ -747,13 +728,20 @@ class MongoDBFileService @Inject() (
         }
 
         // finally delete the actual file
-        ByteStorageService.delete(file.loader, file.loader_id, FileDAO.COLLECTION)
+        if(isLastPointingToLoader(file.loader, file.loader_id)) {
+          ByteStorageService.delete(file.loader, file.loader_id, FileDAO.COLLECTION)
+        }
+
         FileDAO.remove(file)
       }
       case None => Logger.debug("File not found")
     }
   }
 
+  def isLastPointingToLoader(loader: String, loader_id: String): Boolean = {
+    val result = FileDAO.find(MongoDBObject("loader" -> loader, "loader_id" -> loader_id))
+    result.size == 1
+  }
   def removeTemporaries(){
     val cal = Calendar.getInstance()
     val timeDiff = play.Play.application().configuration().getInt("rdfTempCleanup.removeAfter")
@@ -762,7 +750,7 @@ class MongoDBFileService @Inject() (
 
     val tmpDir = System.getProperty("java.io.tmpdir")
     val filesep = System.getProperty("file.separator")
-    val rdfTmpDir = new java.io.File(tmpDir + filesep + "medici__rdfdumptemporaryfiles")
+    val rdfTmpDir = new java.io.File(tmpDir + filesep + "clowder__rdfdumptemporaryfiles")
     if(!rdfTmpDir.exists()){
       rdfTmpDir.mkdir()
     }
@@ -1000,7 +988,7 @@ class MongoDBFileService @Inject() (
 					      mdMoveFile.getParentFile().mkdirs()
 
 						  if(mdFile.renameTo(mdMoveFile)){
-			            	Logger.info("File metadata dumped and moved to staging directory successfully.")
+			            	Logger.debug("File metadata dumped and moved to staging directory successfully.")
 						  }else{
 			            	Logger.warn("Could not move dumped file metadata to staging directory.")
 			            	throw new Exception("Could not move dumped file metadata to staging directory.")
